edit Cargo.toml with feature flags and model loader
