+++ bert-burn/README.md
@@ -18,18 +18,22 @@ bert-burn = { git = "https://github.com/burn-rs/models", package = "bert-burn", 
 
 Example usage for getting sentence embedding from given input text. The model supports multiple backends from burn
 (e.g. `ndarray`, `wgpu`, `tch-gpu`, `tch-cpu`) which can be selected using the `--features` flag. An example with `wgpu`
-backend is shown below.
+backend is shown below. The `fusion` flag is used to enable kernel fusion for the `wgpu` backend. It is not required
+with other backends. The `safetensors` flag is used to support loading `candle-core` safetensors. It will eventually
+be replaced by `burn-safetensors` once it is available.
 
 ### WGPU backend
 
 ```bash
 cd bert-burn/
 # Get sentence embeddings from the RobBERTa encoder (default)
-cargo run --example infer-embedding --release --features wgpu
+cargo run --example infer-embedding --release --features wgpu,fusion,safetensors
 
 # Using bert-base-uncased model
-cargo run --example infer-embedding --release --features wgpu bert-base-uncased 
+cargo run --example infer-embedding --release --features wgpu,fusion,safetensors bert-base-uncased 
 
 # Using roberta-large model
-cargo run --example infer-embedding --release --features wgpu roberta-large
+cargo run --example infer-embedding --release --features wgpu,fusion,safetensors roberta-large
 ```
+
+
